Add snackbar in EvaluateGrades
